test(tooling): split the vue parity runner baseline into its own lane
